build(deps): Move to bundlesize2 (temporarily) to avoid iltorb issues
It seems in the infinite wisdom of NPM ecosystem's churn, the NPM module
bundlesize has chosen to upgrade to a breaking v2 by creating an
entirely new package. This package fixes a few issues with the old one,
including dependency on a long-broken `iltorb` transitive dependency. We
need to move away from iltorb on Torus because of recurring build issues
and its own dependence on node-gyp.

The maintainers of bundlesize2 note that it'll be _eventually_ merged
back into bundlesize as bundlesize@2.0.0, but until then this will have
to do.